fix: moved back a fixture to main conftest, will be removed later once everything was properly moved
