+++ tests/unit_test/conftest.py
@@ -24,17 +24,3 @@ def mock_authenticated_user():
     app.dependency_overrides.pop(get_current_user, None)
 
 
-TOKEN_ENCRYPTED_1 = "gAAAAABoMFiNIvAc7WIFnoKXBjkpAVrdiTFrhlmZtG8BBwvmy1dtvfEFmupm0fcvDUo3unosoAQz5eclP2QFMnPMLG4Hj21MBt-xTdWL661JnWP-wQarnLI="
-
-
-@pytest.fixture
-def mock_encryption_helper():
-    """Mock EncryptionHelper for token encryption/decryption."""
-    with patch("app.routes.git_tokens.EncryptionHelper") as mock_helper:
-        mock_instance = MagicMock()
-        mock_instance.encrypt.return_value = TOKEN_ENCRYPTED_1
-        mock_instance.decrypt.return_value = "ghp_1234567890abcdef"
-        mock_helper.return_value = mock_instance
-        mock_helper.encrypt = mock_instance.encrypt
-        mock_helper.decrypt = mock_instance.decrypt
-        yield mock_helper
